+++ src/main/java/com/publichealthnonprofit/programfunding/controller/model/GrantingOrgData.java
@@ -36,7 +36,7 @@ public class GrantingOrgData {
     public static class GrantingOrgFinancialGrant {
         private Long financialGrantId;
         private String financialGrantName;
-        private Double financialGrantAmount;
+        private Float financialGrantAmount;
         private Date financialGrantStartDate;
         private Date financialGrantEndDate;
